Styled back button. Add moved project code to its own project.component to clean up projectPage.component
